DAOS-647 common: Add method to parse ACL principal
Commands to delete ACL entries need a format that
designates user and group principal names on the
command line.

This patch adds a common method to decode principal
names in that string format to a type and name,
which can be used to manipulate daos_acl and daos_ace
structures.

Signed-off-by: Kris Jacque <kristin.jacque@intel.com>